split nwb gui and data module from visprotocol legacy data format versions
